GH-14: Move JSONError to utils
